#3,The configuration files have been moved from QString to QVariants and system specification that was real numbers now are QVariantList of integers, this fix a bug with validation of system specifications

+++ test/unit/tst_multios.cpp
@@ -7,18 +7,18 @@ void MultiOsTest::initTestCase(){
 }
 
 void MultiOsTest::getProcessorCoresNumberReturnAValidValue(){
-  QVERIFY2(operativeSystem->getCoresNum() > 0, "should be true");
+  QVERIFY2(operativeSystem->getCoresNum().toInt() > 0, "should be true");
 }
 
 void MultiOsTest::getProcessorSseVersionReturnAValidValue(){
-  QVERIFY2(operativeSystem->getSseVersion()>0, "should be true");
+  QVERIFY2(operativeSystem->getSseVersion().toInt()>0, "should be true");
 }
 
 void MultiOsTest::getScreenHeightReturnAValidValue(){
-  QVERIFY2(operativeSystem->getScreenHeight()>0, "should be true");
+  QVERIFY2(operativeSystem->getScreenHeight().toInt()>0, "should be true");
 }
 
 void MultiOsTest::getScreenWidthReturnAValidValue(){
-  QVERIFY2(operativeSystem->getScreenWidth()>0, "should be true");
+  QVERIFY2(operativeSystem->getScreenWidth().toInt()>0, "should be true");
 }
